內鬼群問答 正式/requirements.txt
aiohttp==3.7.4.post0
discord.py==2.0.1